np.int -> np.int32

+++ map/views.py
@@ -1284,8 +1284,8 @@ class MapLayer(object):
 
                 # Check for pixel overlap area (projecting target WCS edges into this brick)
                 ok,xx,yy = bwcs.radec2pixelxy(target_ra, target_dec)
-                xx = xx.astype(np.int)
-                yy = yy.astype(np.int)
+                xx = xx.astype(np.int32)
+                yy = yy.astype(np.int32)
 
                 #print('Brick', brickname, 'band', band, 'shape', bwcs.shape, 'pixel coords', xx, yy)
 
@@ -5011,8 +5011,8 @@ class OutliersLayer(DecalsLayer):
                 continue
             # Check for pixel overlap area (projecting target WCS edges into this brick)
             ok,xx,yy = bwcs.radec2pixelxy(target_ra, target_dec)
-            xx = xx.astype(np.int)
-            yy = yy.astype(np.int)
+            xx = xx.astype(np.int32)
+            yy = yy.astype(np.int32)
             imW,imH = int(bwcs.get_width()), int(bwcs.get_height())
             M = 10
             xlo = np.clip(xx.min() - M, 0, imW)
